Added model layer for furniture
