+++ app/src/main/java/com/example/hsh/homesweethome/furniture/model/Interfaces/IFurnitureCategoryListModel.java
@@ -1,0 +1,16 @@
+package com.example.hsh.homesweethome.furniture.model.Interfaces;
+
+import com.example.hsh.homesweethome.Models.CategoryFurniture;
+
+import java.util.ArrayList;
+
+public interface IFurnitureCategoryListModel {
+
+    interface OnFinishedListener {
+        void onFinished(ArrayList<CategoryFurniture> categories);
+        void onFailure(Throwable t);
+    }
+
+    void getFurnitureList(OnFinishedListener onFinishedListener);
+
+}
