+++ app/src/main/java/com/example/hsh/homesweethome/furniture/presenter/FurnitureMainFragmentPresenter.java
@@ -1,69 +1,41 @@
 package com.example.hsh.homesweethome.furniture.presenter;
 
 import com.example.hsh.homesweethome.Models.CategoryFurniture;
-import com.example.hsh.homesweethome.Models.Furniture;
+import com.example.hsh.homesweethome.furniture.model.FurnitureCategoryCategoryListModel;
+import com.example.hsh.homesweethome.furniture.model.Interfaces.IFurnitureCategoryListModel;
 import com.example.hsh.homesweethome.furniture.presenter.interfaces.IFurnitureMainFragmentPresenter;
 import com.example.hsh.homesweethome.furniture.view.interfaces.IFurnitureMainFragmentView;
-import com.example.hsh.homesweethome.network.APIService;
-import com.example.hsh.homesweethome.network.Utils;
 
 import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
 
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
-
-public class FurnitureMainFragmentPresenter implements IFurnitureMainFragmentPresenter {
+public class FurnitureMainFragmentPresenter implements IFurnitureMainFragmentPresenter, IFurnitureCategoryListModel.OnFinishedListener {
 
     private IFurnitureMainFragmentView view;
-    private APIService mAPIService;
+    private IFurnitureCategoryListModel furnitureCategoryListModel;
 
-    private ArrayList<CategoryFurniture> categories = new ArrayList<>();
 
     public FurnitureMainFragmentPresenter(IFurnitureMainFragmentView view) {
         this.view = view;
+        this.furnitureCategoryListModel = new FurnitureCategoryCategoryListModel();
     }
 
     @Override
     public void getFurniture() {
-
-        mAPIService = Utils.getAPIService();
-        Call<List<Furniture>> call = mAPIService.getFurnitures();
-        call.enqueue(new Callback<List<Furniture>>() {
-            @Override
-            public void onResponse(Call<List<Furniture>> call, Response<List<Furniture>> response) {
-                if (response.isSuccessful()) {
-                    HashMap<String, ArrayList<Furniture>> splitFurnitures = new HashMap<>();
-                    for(Furniture furniture : response.body()) {
-                        if (splitFurnitures.get(furniture.getFurnitureCategory()) != null) {
-                            splitFurnitures.get(furniture.getFurnitureCategory()).add(furniture);
-                        } else {
-                            ArrayList<Furniture> furnitures = new ArrayList<>();
-                            furnitures.add(furniture);
-                            splitFurnitures.put(furniture.getFurnitureCategory(), furnitures);
-                        }
-                    }
-
-                    for (String furnitureCategory : splitFurnitures.keySet()) {
-                        categories.add(new CategoryFurniture(furnitureCategory, splitFurnitures.get(furnitureCategory)));
-                    }
-
-                }
-            }
-
-            @Override
-            public void onFailure(Call<List<Furniture>> call, Throwable t) {
-
-            }
-        });
-
-        view.displayFurnitureRecyclerView(categories);
+        furnitureCategoryListModel.getFurnitureList(this);
     }
 
     @Override
     public void getFilteredFurniture(String query) {
+
+    }
+
+    @Override
+    public void onFinished(ArrayList<CategoryFurniture> categories) {
+        view.displayFurnitureRecyclerView(categories);
+    }
+
+    @Override
+    public void onFailure(Throwable t) {
 
     }
 }
